+++ lib/src/text_overflow_widget.dart
@@ -12,6 +12,17 @@ enum TextOverflowAlign {
 
   /// Align the [TextOverflowWidget] on the right edge of the Text Overflow Rect.
   right,
+
+  /// Align the [TextOverflowWidget] on the center of the Text Overflow Rect.
+  center,
+}
+
+/// The position which TextOverflowWidget should be shown
+/// https://github.com/flutter/flutter/issues/45336
+enum TextOverflowPosition {
+  start,
+  middle,
+  end,
 }
 
 class TextOverflowWidget extends StatelessWidget {
@@ -19,7 +30,8 @@ class TextOverflowWidget extends StatelessWidget {
     required this.child,
     this.align = TextOverflowAlign.right,
     this.maxHeight,
-    this.fixedOffset = Offset.zero,
+    this.position = TextOverflowPosition.end,
+    this.debugOverflowRectColor,
   });
 
   /// The widget of TextOverflow.
@@ -31,8 +43,12 @@ class TextOverflowWidget extends StatelessWidget {
   /// The maxHeight of [TextOverflowWidget], default is preferredLineHeight.
   final double? maxHeight;
 
-  /// Fixed offset refer to the Text Overflow Rect and [child].
-  final Offset fixedOffset;
+  /// The position which TextOverflowWidget should be shown
+  final TextOverflowPosition position;
+
+  /// Whether paint overflow rect, just for debug
+  /// https://github.com/flutter/flutter/issues/45336
+  final Color? debugOverflowRectColor;
 
   @override
   Widget build(BuildContext context) {