+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 6.0.0
+
+* Add [TextOverflowWidget.position] to support show overflow at start, middle or end.
+  https://github.com/flutter/flutter/issues/45336
+* Add [ExtendedText.joinZeroWidthSpace] to make line breaking and overflow style better.
+  https://github.com/flutter/flutter/issues/18761
+* Fix strutStyle not work.
+* Breaking change: remove [TextOverflowWidget.fixedOffset]
+* Breaking change: [SpecialText.getContent] is not include endflag now.(please check if you call getContent and your endflag length is more than 1)
+
 ## 5.0.5
 
 * Fix issue that childIndex == children.length assert false in assembleSemanticsNode when use overflowWidget and text is not overflow.
@@ -5,7 +15,7 @@
 ## 5.0.4
 
 * Fix issue that the overflowWidget is not layout #97
-  
+
 ## 5.0.3
 
 * Fix null-safety error #96
@@ -28,13 +38,13 @@
 
 * Support keyboard copy on web/desktop
 * Fix wrong position of caret
-  
+
 ## 4.0.1
 
 * Change handleSpecialText to hasSpecialInlineSpanBase(extended_text_library)
 * Add hasPlaceholderSpan(extended_text_library)
 * Fix wrong offset of WidgetSpan #86
-  
+
 ## 4.0.0
 
 * Merge from Flutter v1.20